+++ database/migrations/2019_03_15_165709_create_comments_table.php
@@ -14,12 +14,12 @@ class CreateCommentsTable extends Migration
     public function up()
     {
         Schema::create('comments', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->text('comment');
-            $table->unsignedInteger('parent_id')->nullable();
+            $table->unsignedBigInteger('parent_id')->nullable();
 
-            $table->unsignedInteger('created_by');
-            $table->unsignedInteger('contact_id');
+            $table->unsignedBigInteger('created_by');
+            $table->unsignedBigInteger('contact_id');
 
             $table->timestamps();
