+++ database/migrations/2019_03_25_113236_create_permission_tables.php
@@ -16,8 +16,8 @@ class CreatePermissionTables extends Migration
         $tableNames = config('permission.table_names');
 
         Schema::create($tableNames['roles'], function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('team_id')->nullable();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('team_id')->nullable();
             $table->string('name');
             $table->string('guard_name');
             $table->string('slug');
@@ -31,7 +31,7 @@ class CreatePermissionTables extends Migration
         });
 
         Schema::create($tableNames['model_has_permissions'], function (Blueprint $table) use ($tableNames) {
-            $table->unsignedInteger('permission_id');
+            $table->unsignedBigInteger('permission_id');
             $table->morphs('model');
 
             $table->foreign('permission_id')
@@ -43,7 +43,7 @@ class CreatePermissionTables extends Migration
         });
 
         Schema::create($tableNames['model_has_roles'], function (Blueprint $table) use ($tableNames) {
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->morphs('model');
 
             $table->foreign('role_id')
@@ -55,8 +55,8 @@ class CreatePermissionTables extends Migration
         });
 
         Schema::create($tableNames['role_has_permissions'], function (Blueprint $table) use ($tableNames) {
-            $table->unsignedInteger('permission_id');
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('permission_id');
+            $table->unsignedBigInteger('role_id');
 
             $table->foreign('permission_id')
                 ->references('id')
