+++ database/migrations/2019_06_23_085048_create_announcement_user_table.php
@@ -14,8 +14,8 @@ class CreateAnnouncementUserTable extends Migration
     public function up()
     {
         Schema::create('announcement_user', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('announcement_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('announcement_id');
             $table->timestamps();
 
             $table->foreign('user_id')
